requirements.txt
APScheduler==3.10.4
asgiref==3.8.1
bcrypt==4.1.2
Django==4.1.13
django-apscheduler==0.6.2
django-redis==5.4.0
djongo==1.3.6
dnspython==2.6.1
PyJWT==2.8.0
pymongo==3.12.3
PyMySQL==1.1.0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.3
six==1.16.0
sqlparse==0.2.4
tzlocal==5.2
